+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.8.0 (2025-02-21)
+
+### Feat
+
+- add dir to gitignore
+
+## 0.7.1 (2025-02-21)
+
+### Fix
+
+- Fix cn and alt_names order
+- Fix list certificate with filters expired-since and expiring-in
+
 ## 0.7.0 (2025-02-21)
 
 ### Feat